handle the case of a unicode argument to loads() differently
unfortunately this just needs to be inlined, a separate utf8_z_hash
function won't work. the reason being that PyString_AsStringAndSize
gives access to the string object's internal buffer rather than copying,
and that string object was being decrefed away at the end of the
function, but the buffer we got out of it still needed to be usable.